Game Ready

+++ app/src/main/kotlin/com/shiva0705/sample/musicfy/models/Song.kt
@@ -2,7 +2,7 @@ package com.shiva0705.sample.musicfy.models
 
 class Song(
         val name: String,
-        val popularity: Double,
+        val popularity: Int,
         val type: String,
         val uri: String,
         val artists : List<Artist>,
